abstract placeholder
